Force input directories to absolute paths
Fix failure to refresh if cwd changes when diffing relative paths.

Ensure we're tracking absolute paths so if cwd changes, we don't lose
the correct path.

This looks like a lot of change, but it's just removing the a: and
adding the two fnamemodify lines.

Test Case
    :cd ~/.vim/bundle/zfdirdiff/
    :ZFDirDiff ./ ../test/
    :cd /
    :norm DD

(DD is the default update mapping.)